+++ tests/integration/api/classification/test_api_classification.py
@@ -12,10 +12,7 @@ import numpy as np
 import pytest
 from bson import ObjectId
 
-from otx.algorithms.classification.tasks import (
-    ClassificationInferenceTask,
-    ClassificationTrainTask,
-)
+from otx.algorithms.classification.adapters.mmcls.task import MMClassificationTask
 from otx.algorithms.common.tasks.training_base import BaseTask
 from otx.api.configuration.helper import create
 from otx.api.entities.annotation import (
@@ -203,7 +200,7 @@ class TestClassificationTaskAPI(ClassificationTaskAPIBase):
         task_environment, dataset = self.init_environment(
             hyper_parameters, model_template, multilabel, hierarchical, 20
         )
-        task = ClassificationTrainTask(task_environment=task_environment)
+        task = MMClassificationTask(task_environment=task_environment)
         print("Task initialized, model training starts.")
 
         training_progress_curve = []
@@ -233,7 +230,7 @@ class TestClassificationTaskAPI(ClassificationTaskAPIBase):
         task_environment, dataset = self.init_environment(
             hyper_parameters, model_template, multilabel, hierarchical, 20
         )
-        task = ClassificationInferenceTask(task_environment=task_environment)
+        task = MMClassificationTask(task_environment=task_environment)
         print("Task initialized, model inference starts.")
 
         inference_progress_curve = []
@@ -262,7 +259,7 @@ class TestClassificationTaskAPI(ClassificationTaskAPIBase):
         )
         val_dataset = dataset.get_subset(Subset.VALIDATION)
 
-        train_task = ClassificationTrainTask(task_environment=classification_environment)
+        train_task = MMClassificationTask(task_environment=classification_environment)
 
         training_progress_curve = []
 
@@ -280,7 +277,7 @@ class TestClassificationTaskAPI(ClassificationTaskAPIBase):
 
         # Create InferenceTask
         classification_environment.model = trained_model
-        inference_task = ClassificationInferenceTask(task_environment=classification_environment)
+        inference_task = MMClassificationTask(task_environment=classification_environment)
 
         performance_after_load = task_eval(inference_task, trained_model, val_dataset)